MS-65: refactor Layout folder inside Component folder
